refactor: typescript db
