+++ .github/workflows/ci.yml
@@ -27,7 +27,7 @@ jobs:
         with:
           find: "const BASE_PATH = '/hsr-optimizer'"
           replace: "const BASE_PATH = '/dreary-quibbles'"
-          include: "src/lib/db.js"
+          include: "src/lib/db.ts"
       - name: Find and Replace API endpoint
         uses: jacobtomlinson/gha-find-replace@v3
         with:
