+++ src/components/BuildsModal.tsx
@@ -4,7 +4,7 @@ import { Button, Card, Flex, Modal } from 'antd'
 import StatText from './characterPreview/StatText'
 import { HeaderText } from 'components/HeaderText'
 import { DeleteOutlined, ExclamationCircleOutlined } from '@ant-design/icons'
-import DB from '../lib/db'
+import DB from 'lib/db'
 import { SaveState } from 'lib/saveState'
 import { Message } from 'lib/message'
 import { Character, SavedBuild } from 'types/Character'
