+++ src/components/RelicFilterBar.jsx
@@ -4,7 +4,7 @@ import { RelicScorer } from 'lib/relicScorerPotential'
 import CheckableTag from 'antd/lib/tag/CheckableTag'
 import { HeaderText } from 'components/HeaderText'
 import { TooltipImage } from './TooltipImage'
-import DB from '../lib/db'
+import DB from 'lib/db'
 import { Hint } from 'lib/hint'
 import { Utils } from 'lib/utils'
 import { Constants, SetsRelics, setToId, Stats, UnreleasedSets } from 'lib/constants'
