+++ src/components/RelicsTab.jsx
@@ -11,7 +11,7 @@ import { Gradient } from 'lib/gradient'
 import { Message } from 'lib/message'
 import { TooltipImage } from './TooltipImage'
 import RelicFilterBar from './RelicFilterBar'
-import DB from '../lib/db'
+import DB from 'lib/db'
 import { Assets } from 'lib/assets'
 import { Renderer } from 'lib/renderer'
 import { SaveState } from 'lib/saveState'
