+++ src/components/optimizerTab/OptimizerSuggestionsModal.tsx
@@ -3,7 +3,7 @@ import { ReactElement } from 'react'
 import { Optimizer } from 'lib/optimizer/optimizer'
 import { HorizontalDivider } from 'components/Dividers'
 import { Parts } from 'lib/constants'
-import DB, { AppPages } from 'lib/db.js'
+import DB, { AppPages } from 'lib/db'
 import { Message } from 'lib/message.js'
 import { OptimizerTabController } from 'lib/optimizerTabController.js'
 import { Utils } from 'lib/utils.js'
