+++ src/components/optimizerTab/optimizerForm/OptimizerOptionsDisplay.tsx
@@ -6,7 +6,7 @@ import { Hint } from 'lib/hint.jsx'
 import { HeaderText } from 'components/HeaderText'
 import { TooltipImage } from 'components/TooltipImage.jsx'
 import { useMemo } from 'react'
-import DB from 'lib/db.js'
+import DB from 'lib/db'
 import { optimizerTabDefaultGap, panelWidth } from 'components/optimizerTab/optimizerTabConstants'
 import { Utils } from 'lib/utils.js'
 import { Assets } from 'lib/assets'
