+++ src/components/optimizerTab/optimizerForm/OrnamentSetTagRenderer.tsx
@@ -4,7 +4,7 @@ import { Constants } from 'lib/constants'
 import PropTypes from 'prop-types'
 import React from 'react'
 
-// NOTE: Be careful hot-reloading with this file, can cause DB to wipe. Unsure why yet
+// NOTE: Be careful hot-reloading with this file, can cause Db to wipe. Unsure why yet
 export function OrnamentSetTagRenderer(props) {
   const { value, closable, onClose } = props
   const onPreventMouseDown = (event) => {
