+++ src/components/optimizerTab/optimizerForm/TeammateCard.tsx
@@ -9,7 +9,7 @@ import { LightConeConditionals } from 'lib/lightConeConditionals.js'
 import { OptimizerTabController } from 'lib/optimizerTabController.js'
 import { CharacterConditionalDisplay } from 'components/optimizerTab/conditionals/CharacterConditionalDisplay'
 import { LightConeConditionalDisplay } from 'components/optimizerTab/conditionals/LightConeConditionalDisplay'
-import DB from 'lib/db.js'
+import DB from 'lib/db'
 import { Character } from 'types/Character'
 import { Message } from 'lib/message.js'
 import LightConeSelect from 'components/optimizerTab/optimizerForm/LightConeSelect'
