+++ src/lib/relicRollGrader.test.ts
@@ -2,7 +2,7 @@ import { expect, test } from 'vitest'
 
 import { Constants } from 'lib/constants'
 import { RelicRollGrader } from './relicRollGrader.js'
-import DB from './db.js'
+import DB from 'lib/db'
 import { Relic } from 'types/Relic'
 import { DataParser } from './dataParser.js'
