+++ src/window.d.ts
@@ -3,7 +3,7 @@ import { Constants } from './lib/constants'
 import { OcrParserKelz3 } from 'lib/ocrParserKelz3.jsx'
 import { DataParser } from './lib/dataParser'
 import { OptimizerTabController } from './lib/optimizerTabController'
-import { DB } from './lib/db'
+import { DB } from 'lib/db'
 import { CharacterStats } from './lib/characterStats'
 import { Utils } from './lib/utils'
 import { Assets } from './lib/assets'
